css/css-initial-letter/initial-letter-valid.html

<!DOCTYPE html>
<title>Tests parsing of the initial-letter property</title>
<link rel="help" href="https://https://w3c.github.io/csswg-drafts/css-inline-3/#sizing-drop-initials">
<link rel="author" href="mailto:yosin@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value('initial-letter', 'foobar');
test_invalid_value('initial-letter', 'drop');
test_invalid_value('initial-letter', 'raise');
test_invalid_value('initial-letter', 'drop 1 2');
test_invalid_value('initial-letter', 'raise 1 2');
test_invalid_value('initial-letter', '1 2 3');
test_invalid_value('initial-letter', '1 drop 3');
test_invalid_value('initial-letter', '1 raise 3');
test_invalid_value('initial-letter', '1 foobar 3');

test_valid_value('initial-letter', 'normal');
test_valid_value('initial-letter', '1.23');
test_valid_value('initial-letter', '1.23 456');
test_valid_value('initial-letter', '1.23 drop');
test_valid_value('initial-letter', '1.23 raise');
test_valid_value('initial-letter', 'drop 1.23', '1.23 drop');
test_valid_value('initial-letter', 'raise 1.23', '1.23 raise');
</script>
